# backend/ms-cv/src/main/java/uasz/alumni/ms_cv/controllers/LanguesApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package uasz.alumni.ms_cv.controllers;

import uasz.alumni.ms_cv.models.Langue;
import uasz.alumni.ms_cv.models.LangueCreate;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-26T19:36:29.996620853Z[Africa/Dakar]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "Langues", description = "the Langues API")
public interface LanguesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_LANGUES_ID_DELETE = "/langues/{id}";
    /**
     * DELETE /langues/{id} : Supprimer une langue
     *
     * @param id  (required)
     * @return Langue supprimée (status code 204)
     */
    @Operation(
        operationId = "languesIdDelete",
        summary = "Supprimer une langue",
        tags = { "Langues" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Langue supprimée")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = LanguesApi.PATH_LANGUES_ID_DELETE
    )
    default ResponseEntity<Void> languesIdDelete(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_LANGUES_ID_PUT = "/langues/{id}";
    /**
     * PUT /langues/{id} : Modifier une langue
     *
     * @param id  (required)
     * @param langueCreate  (required)
     * @return Langue mise à jour (status code 200)
     */
    @Operation(
        operationId = "languesIdPut",
        summary = "Modifier une langue",
        tags = { "Langues" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Langue mise à jour")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = LanguesApi.PATH_LANGUES_ID_PUT,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> languesIdPut(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "LangueCreate", description = "", required = true) @Valid @RequestBody LangueCreate langueCreate
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_LANGUES_POST = "/langues";
    /**
     * POST /langues : Ajouter une langue
     *
     * @param langueCreate  (required)
     * @return Langue ajoutée (status code 201)
     */
    @Operation(
        operationId = "languesPost",
        summary = "Ajouter une langue",
        tags = { "Langues" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Langue ajoutée", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Langue.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = LanguesApi.PATH_LANGUES_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Langue> languesPost(
        @Parameter(name = "LangueCreate", description = "", required = true) @Valid @RequestBody LangueCreate langueCreate
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"cvId\" : 1, \"id\" : 1, \"nom\" : \"Français\", \"niveau\" : \"LANGUE_MATERNELLE\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}