# openapi/ms-cv.yaml
openapi: 3.0.3
info:
  title: CV Service API
  version: 1.0.0
  description: API de gestion des CV des alumni et étudiants - Génération de CV standardisés
servers:
  - url: http://localhost:8082/api/v1/cv
paths:
...
  /langues:
    post:
      tags:
        - Langues
      summary: Ajouter une langue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LangueCreate"
      responses:
        "201":
          description: Langue ajoutée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Langue"
  /langues/{id}:
    put:
      tags:
        - Langues
      summary: Modifier une langue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LangueCreate"
      responses:
        "200":
          description: Langue mise à jour

    delete:
      tags:
        - Langues
      summary: Supprimer une langue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Langue supprimée
...
components:
  schemas:
...
    Langue:
      type: object
      properties:
        id:
          type: integer
          example: 1
        cvId:
          type: integer
          example: 1
        nom:
          type: string
          example: "Français"
        niveau:
          type: string
          enum: [DEBUTANT, INTERMEDIAIRE, AVANCE, COURANT, LANGUE_MATERNELLE]
          example: "LANGUE_MATERNELLE"
    LangueCreate:
      type: object
      required:
        - cvId
        - nom
        - niveau
      properties:
        cvId:
          type: integer
        nom:
          type: string
        niveau:
          type: string
          enum: [DEBUTANT, INTERMEDIAIRE, AVANCE, COURANT, LANGUE_MATERNELLE]
...